// src/pages/timezone/types/palette.ts
enum Palette {
  red = '#E32322',
  darkOrange = '#E96220',
  orange = '#F28E1C',
  lightOrange = '#FDC60A',
  yellow = '#E5DD00',
  lightGreen = '#8DBB25',
  green = '#008F5A',
  lightblue = '#0696BB',
  blue = '#2A71AF',
  purple = '#454E99',
  darkPurple = '#6D398B',
  lightPurple = '#C5037D',
  gray = '#8C8C8C',
}

export default Palette;